The sess-id field of the o= line can't be 64 bits
RFC 3264 says this:

    The numeric value of the session id and version in the o line MUST
    be representable with a 64 bit signed integer.

Since the negative values that can be represented in a 64-bit signed
integer can't be represented by the sess-id grammar (`1*DIGIT`),
the only reasonable conclusion is that this has to be a 63-bit
number.

FWIW, that's what Firefox does, though it will accept 64-bit values
(apparently longer values clamp to ULLONG_MAX).