Describing when the ICE candidate pool is used and can be resized.
Fixes #381

The pool is only used in the first offer/answer exchange. After applying
a local description, the pool size is frozen and changing the ICE
servers will *not* result in new candidates being pooled. After an answer
is applied, any leftover candidates are discarded. This only happens
when applying an answer so that candidates aren't prematurely discarded
for use cases that involve pranswer, or multiple offers.