Proposal for updated RTP/RTCP Muxing proposal (Appendix B)
This updated proposal builds on the old Appendix B. And has mostly the
same behavior but describes things using RTP streams, and attempts to
makes no assumptions of how the forwarding of the information to the
relevant consumer or handler is located in relation to the rest of the
RTP/RTCP processing. It is clear on that PT demuxing can't be used other
than initial binding of a new SSRC. It also clarifies RTCP third party
reporting into a seperate category thus reducing overhead in duplicating
infromation form an algorithmic level. Handling of third party
monitoriing is now much more implementation specific.