Fixing some text that deals with currentDirection.
Fixes #757.

`currentDirection` is meant to represent "direction in effect", whereas
`direction` represents "direction the application *wants* to negotiate".
So, the direction that `removeTrack` both inspects and modifies should
be `direction`; its purpose is to change the application's desired
direction for a transceiver, regardless of the current state of
negotiation.

Also clarifying that `currentDirection` also reflects the direction
negotiated in a provisional answer, which was the intent (I believe)
but wasn't explicitly stated. This is consistent with it being the
direction "in effect", since applying a provisional answer allows the
endpoint to start sending/receiving media.